Access/requirement.txt
Django>=5.0
mysqlclient>=2.0
python-decouple
pyjwt